incorporate some of paul's feedback

+++ Sources/ApodiniUtils/FileManager.swift
@@ -9,31 +9,31 @@ import Foundation
 
 
 extension FileManager {
-    public func lk_initialize() throws {
-        try createDirectory(at: lk_temporaryDirectory, withIntermediateDirectories: true, attributes: [:])
+    public func initialize() throws {
+        try createDirectory(at: apodiniDeployTmpDir, withIntermediateDirectories: true, attributes: [:])
     }
     
     
-    public func lk_directoryExists(atUrl url: URL) -> Bool {
+    public func directoryExists(atUrl url: URL) -> Bool {
         var isDirectory: ObjCBool = false
         return fileExists(atPath: url.path, isDirectory: &isDirectory) && isDirectory.boolValue
     }
     
     
-    public func lk_setWorkingDirectory(to newDir: URL) throws {
+    public func setWorkingDirectory(to newDir: URL) throws {
         guard changeCurrentDirectoryPath(newDir.path) else {
             throw ApodiniUtilsError(message: "Unable to change working directory from \(self.currentDirectoryPath) to \(newDir.path)")
         }
     }
     
     
-    public var lk_temporaryDirectory: URL {
-        temporaryDirectory.appendingPathComponent("Apodini", isDirectory: true)
+    public var apodiniDeployTmpDir: URL {
+        temporaryDirectory.appendingPathComponent("ApodiniDeploy", isDirectory: true)
     }
     
     
-    public func lk_getTemporaryFileUrl(fileExtension: String?) -> URL {
-        var tmpfile = lk_temporaryDirectory
+    public func getTemporaryFileUrl(fileExtension: String?) -> URL {
+        var tmpfile = apodiniDeployTmpDir
             .appendingPathComponent(UUID().uuidString)
         if let ext = fileExtension {
             tmpfile.appendPathExtension(ext)
@@ -42,9 +42,8 @@ extension FileManager {
     }
     
     
-    public func lk_copyItem(at srcUrl: URL, to dstUrl: URL) throws {
-        // TODO look into the -replace API
-        if fileExists(atPath: dstUrl.path) {
+    public func copyItem(at srcUrl: URL, to dstUrl: URL, overwriteExisting: Bool) throws {
+        if overwriteExisting && fileExists(atPath: dstUrl.path) {
             try removeItem(at: dstUrl)
         }
         try copyItem(at: srcUrl, to: dstUrl)
@@ -56,7 +55,7 @@ extension FileManager {
 
 extension FileManager {
     /// Read file permissions
-    public func lk_posixPermissions(ofItemAt url: URL) throws -> POSIXPermissions {
+    public func posixPermissions(ofItemAt url: URL) throws -> POSIXPermissions {
         if let value = try self.attributesOfItem(atPath: url.absoluteURL.path)[.posixPermissions] as? NSNumber {
             return POSIXPermissions(rawValue: numericCast(value.uintValue))
         } else {
@@ -67,7 +66,7 @@ extension FileManager {
     }
     
     /// Write file permissions
-    public func lk_setPosixPermissions(_ permissions: POSIXPermissions, forItemAt url: URL) throws {
+    public func setPosixPermissions(_ permissions: POSIXPermissions, forItemAt url: URL) throws {
         try url.withUnsafeFileSystemRepresentation { ptr in
             guard let ptr = ptr else {
                 throw ApodiniUtilsError(message: "Unable to set file permissins: can't get file path")
